requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
pydantic[email]
pyjwt
python-dotenv
alembic
passlib
bcrypt